Move query_yes_no and rename url
