+++ mikecloudio/timeseries.py
@@ -1,12 +1,11 @@
 import json
+import sys
 import warnings
 from pathlib import Path
 # import matplotlib.pyplot as plt
 
 import pandas as pd
 import requests
-
-from mikecloudio.request import query_yes_no
 
 
 class Dataset:
@@ -559,3 +558,36 @@ class Timeseries:
         response = requests.delete(url, headers=self._header)
         if response.status_code > 300:
             raise ValueError("request failed. make sure times are in format {yyyy-MM-ddTHHmmss}")
+
+
+def query_yes_no(question, default="yes"):
+    """Ask a yes/no question via raw_input() and return their answer.
+
+    "question" is a string that is presented to the user.
+    "default" is the presumed answer if the user just hits <Enter>.
+        It must be "yes" (the default), "no" or None (meaning
+        an answer is required of the user).
+
+    The "answer" return value is True for "yes" or False for "no".
+    """
+    valid = {"yes": True, "y": True, "ye": True,
+             "no": False, "n": False}
+    if default is None:
+        prompt = " [y/n] "
+    elif default == "yes":
+        prompt = " [Y/n] "
+    elif default == "no":
+        prompt = " [y/N] "
+    else:
+        raise ValueError("invalid default answer: '%s'" % default)
+
+    while True:
+        sys.stdout.write(question + prompt)
+        choice = input().lower()
+        if default is not None and choice == '':
+            return valid[default]
+        elif choice in valid:
+            return valid[choice]
+        else:
+            sys.stdout.write("Please respond with 'yes' or 'no' "
+                             "(or 'y' or 'n').\n")
